refactor: move IncomingEncryptSubmission to own file
