+++ src/app/modules/form/admin-form/admin-form.controller.ts
@@ -62,10 +62,8 @@ import {
 } from '../../submission/email-submission/email-submission.util'
 import * as EncryptSubmissionMiddleware from '../../submission/encrypt-submission/encrypt-submission.middleware'
 import * as EncryptSubmissionService from '../../submission/encrypt-submission/encrypt-submission.service'
-import {
-  IncomingEncryptSubmission,
-  mapRouteError as mapEncryptSubmissionError,
-} from '../../submission/encrypt-submission/encrypt-submission.utils'
+import { mapRouteError as mapEncryptSubmissionError } from '../../submission/encrypt-submission/encrypt-submission.utils'
+import IncomingEncryptSubmission from '../../submission/encrypt-submission/IncomingEncryptSubmission.class'
 import * as SubmissionService from '../../submission/submission.service'
 import {
   extractEmailConfirmationData,
